fix: timecop specs
Moved to the crystal-community maintained package, and now using blocks.

It appears that Spec.before_each is a globally scoped call and the load order of files will determine which one is used. Moved to providing a time freeze within a block, and this is recommended by Timecop's safe mode.